# docs/swagger.yaml
basePath: /
definitions:
  async.CalculationResult:
    properties:
      calc_id:
        type: string
      res_en:
        type: number
      res_ni:
        type: number
      star_id:
        type: string
    type: object
  ds.Calc:
    properties:
      inp_dist:
        type: number
      inp_mass:
        type: number
      inp_texp:
        type: number
      res_en:
        type: number
      res_ni:
        type: number
      scope_id:
        type: integer
      star_id:
        type: integer
    type: object
  ds.Scope:
    properties:
      delta_lamb:
        type: number
      description:
        type: string
      filter:
        type: string
      id:
        type: integer
      img_link:
        type: string
      lambda:
        type: number
      name:
        type: string
      status:
        type: boolean
      zero_point:
        type: number
    type: object
...
host: 127.0.0.1
info:
  contact:
    email: Barinovartem383@gmail.com
    name: API Support
    url: https://vk.com/aaaaaaeaaaa
  description: Bmstu Open IT Platform
  license:
    name: AS IS (NO WARRANTY)
  title: Energy Calculation System
  version: "1.0"
paths:
  /calc/:
    put:
      description: Callback endpoint for async service to send calculation results
      parameters:
      - description: Calculation result
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/async.CalculationResult'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Receive calculation result from async service
      tags:
      - Calculation
  /calc/delete/{star_id}/{scope_id}:
    delete:
      description: Delete calculation from user's star. If no calculations remain,
        the star is deleted.
      parameters:
      - description: Star ID
        in: path
        name: star_id
        required: true
        type: string
      - description: Scope ID
        in: path
        name: scope_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete calculation from star
      tags:
      - Calc
  /calc/edit/{star_id}/{scope_id}:
    put:
      description: Update calculation parameters in user's star
      parameters:
      - description: Star ID
        in: path
        name: star_id
        required: true
        type: string
      - description: Scope ID
        in: path
        name: scope_id
        required: true
        type: string
      - description: Calculation parameters
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Calc'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Edit calculation in star
      tags:
      - Calc
  /scope/{id}:
    get:
      description: Get detailed information about specific scope
      parameters:
      - description: Scope ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get scope by ID
      tags:
      - Scope
  /scope/add:
    post:
      description: Create new scope (moderator only)
      parameters:
      - description: Scope data
        in: body
        name: scope
        required: true
        schema:
          $ref: '#/definitions/ds.Scope'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.Scope'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add new scope
      tags:
      - Scope
  /scope/addpicture/{id}:
    post:
      description: Upload image for scope (moderator only)
      parameters:
      - description: Scope ID
        in: path
        name: id
        required: true
        type: string
      - description: Scope image
        in: formData
        name: image
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add scope picture
      tags:
      - Scope
  /scope/addtostar:
    post:
      description: Add scope to current user's active star
      parameters:
      - description: Scope ID
        in: formData
        name: scope_id
        required: true
        type: string
      responses:
        "303":
          description: Redirect to /scopes
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add scope to star
      tags:
      - Scope
  /scope/delete/{id}:
    delete:
      description: Delete scope and its image (moderator only)
      parameters:
      - description: Scope ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete scope
      tags:
      - Scope
  /scope/edit/{id}:
    put:
      description: Update scope information (moderator only)
      parameters:
      - description: Scope ID
        in: path
        name: id
        required: true
        type: string
      - description: Scope data
        in: body
        name: scope
        required: true
        schema:
          $ref: '#/definitions/ds.Scope'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Scope'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Edit scope
      tags:
      - Scope
  /scopes:
    get:
      description: Get all scopes with optional search filter
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get scopes list
      tags:
      - Scope
...
schemes:
- http
swagger: "2.0"